requirements.txt
feedparser
httpx
trafilatura
PyYAML
selectolax
orjson
sentence-transformers
faiss-cpu
scikit-learn
pandas
numpy
openpyxl